requirements.txt
uvicorn[standard]
gunicorn
dotenv
fastapi
pymongo
pyjwt
bcrypt
python-multipart